+++ apps/bare-example/src/Movies.tsx
@@ -1,12 +1,6 @@
 import { Text, View, StyleSheet, ActivityIndicator, FlatList, RefreshControl, SafeAreaView } from 'react-native';
 import { useCallback, useEffect, useState } from 'react';
-import {
-  makeGetRequestWithWebView,
-  MOVIES,
-  StreamingServers,
-  type IMovieResult,
-  type ISearch,
-} from 'react-native-consumet';
+import { MOVIES, type IMovieResult, type ISearch } from 'react-native-consumet';
 
 interface FetchState {
   data: IMovieResult[];
@@ -17,12 +11,12 @@ interface FetchState {
 const fetchData = async (): Promise<ISearch<IMovieResult>> => {
   try {
     const movies = new MOVIES.NetflixMirror();
-    const search = await movies.search('dr. stone');
+    const search = await movies.search('jujutsu kaisen');
     console.log(search);
     const info = await movies.fetchMediaInfo(search.results![0]!.id);
     console.log(info, info.episodes![0]!.id);
 
-    const sources = await movies.fetchEpisodeSources("81144552", "81144552");
+    const sources = await movies.fetchEpisodeSources('81144552', '81144552');
     // const servers = await movies.fetchEpisodeServers(info.episodes![0]!.id);
     console.log(sources);
     // const s = await movies.fetchEpisodeSources(